masih ada error di controller product

+++ entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	ID       uint16 `gorm:"primary_key;auto_increment" json:"id,omitempty"`
+	ID       int `gorm:"primary_key;auto_increment" json:"id,omitempty"`
 	Name     string `json:"name" binding:"required" gorm:"type:varchar(400)"`
 	Email    string `json:"email" binding:"required" gorm:"type:varchar(400);unique"`
 	Password string `json:"password" binding:"required" gorm:"type:varchar(400)"`
@@ -12,4 +12,5 @@ type User struct {
 	Country  string `json:"country" binding:"required" gorm:"type:varchar(400)"`
 	ZipCode  string `json:"zip_code" binding:"required" gorm:"type:varchar(400)"`
 	Role     string `json:"role" binding:"required" gorm:"type:varchar(400)"`
+
 }
